requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2
django-admin==2.0.2
django-environ==0.10.0
django-excel-response2==3.0.5
django-filter==23.1
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
exceptiongroup==1.1.1
iniconfig==2.0.0
isoweek==1.3.3
Markdown==3.4.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tomli==2.0.1
tzdata==2023.3
tzlocal==4.3
xlwt==1.3.0
